backend/pyproject.toml
[project]
name = "resume-analyser"
version = "0.1.0"
description = "Insert description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dateparser>=1.2.1",
    "dynaconf>=3.2.10",
    "fastapi[standard]>=0.115.12",
    "litellm>=1.66.1",
    "loguru>=0.7.3",
    "pillow>=11.2.1",
    "pydantic>=2.11.3",
    "qdrant-client>=1.14.2",
    "requests>=2.32.3",
    "sqlmodel>=0.0.24",
    "typer>=0.15.2",
]

[dependency-groups]
dev = [
    "ruff>=0.11.6",
]
